Adjusted according to review comments
