+++ frontend/src/modules/2DViewer/DataProviderFramework/customDataProviderImplementations/StatisticalSurfaceProvider.ts
@@ -1,11 +1,17 @@
 import type { SurfaceDataPng_api } from "@api";
-import { SurfaceTimeType_api, getRealizationSurfacesMetadataOptions, getSurfaceDataOptions } from "@api";
+import {
+    SurfaceStatisticFunction_api,
+    SurfaceTimeType_api,
+    getRealizationSurfacesMetadataOptions,
+    getSurfaceDataOptions,
+} from "@api";
 import type {
     CustomDataProviderImplementation,
     DataProviderInformationAccessors,
     FetchDataParams,
 } from "@modules/_shared/DataProviderFramework/interfacesAndTypes/customDataProviderImplementation";
 import type { DefineDependenciesArgs } from "@modules/_shared/DataProviderFramework/interfacesAndTypes/customSettingsHandler";
+import type { SensitivityNameCasePair } from "@modules/_shared/DataProviderFramework/settings/implementations/SensitivitySetting";
 import type { MakeSettingTypesMap } from "@modules/_shared/DataProviderFramework/settings/settingsDefinitions";
 import { Setting } from "@modules/_shared/DataProviderFramework/settings/settingsDefinitions";
 import type { FullSurfaceAddress } from "@modules/_shared/Surface";
@@ -16,30 +22,31 @@ import { encodeSurfAddrStr } from "@modules/_shared/Surface/surfaceAddress";
 
 import { isEqual } from "lodash";
 
-const realizationSurfaceSettings = [
+const statisicalSurfaceSettings = [
     Setting.ENSEMBLE,
-    Setting.REALIZATION,
+    Setting.STATISTIC_FUNCTION,
+    Setting.SENSITIVITY,
     Setting.ATTRIBUTE,
     Setting.SURFACE_NAME,
     Setting.TIME_OR_INTERVAL,
     Setting.COLOR_SCALE,
 ] as const;
-export type RealizationSurfaceSettings = typeof realizationSurfaceSettings;
-type SettingsWithTypes = MakeSettingTypesMap<RealizationSurfaceSettings>;
+export type StatisticalSurfaceSettings = typeof statisicalSurfaceSettings;
+type SettingsWithTypes = MakeSettingTypesMap<StatisticalSurfaceSettings>;
 
 export enum SurfaceDataFormat {
     FLOAT = "float",
     PNG = "png",
 }
 
-export type RealizationSurfaceData =
+export type StatisticalSurfaceData =
     | { format: SurfaceDataFormat.FLOAT; surfaceData: SurfaceDataFloat_trans }
     | { format: SurfaceDataFormat.PNG; surfaceData: SurfaceDataPng_api };
 
-export class RealizationSurface
-    implements CustomDataProviderImplementation<RealizationSurfaceSettings, RealizationSurfaceData>
+export class StatisticalSurfaceProvider
+    implements CustomDataProviderImplementation<StatisticalSurfaceSettings, StatisticalSurfaceData>
 {
-    settings = realizationSurfaceSettings;
+    settings = statisicalSurfaceSettings;
 
     private _dataFormat: SurfaceDataFormat;
 
@@ -47,41 +54,23 @@ export class RealizationSurface
         this._dataFormat = dataFormat ?? SurfaceDataFormat.PNG;
     }
 
+    getDefaultSettingsValues() {
+        return {
+            [Setting.STATISTIC_FUNCTION]: SurfaceStatisticFunction_api.MEAN,
+        };
+    }
+
     getDefaultName(): string {
-        return "Realization Surface";
+        return "Statistical Surface";
     }
 
     doSettingsChangesRequireDataRefetch(prevSettings: SettingsWithTypes, newSettings: SettingsWithTypes): boolean {
-        return !isEqual(
-            {
-                ...prevSettings,
-                colorScale: null,
-            },
-            { ...newSettings, colorScale: null },
-        );
-    }
-
-    areCurrentSettingsValid({
-        getSetting,
-    }: DataProviderInformationAccessors<RealizationSurfaceSettings, RealizationSurfaceData>): boolean {
-        const ensembleIdent = getSetting(Setting.ENSEMBLE);
-        const realizationNum = getSetting(Setting.REALIZATION);
-        const surfaceName = getSetting(Setting.SURFACE_NAME);
-        const attribute = getSetting(Setting.ATTRIBUTE);
-        const timeOrInterval = getSetting(Setting.TIME_OR_INTERVAL);
-
-        return (
-            ensembleIdent !== null &&
-            realizationNum !== null &&
-            surfaceName !== null &&
-            attribute !== null &&
-            timeOrInterval !== null
-        );
+        return !isEqual(prevSettings, newSettings);
     }
 
     makeValueRange({
         getData,
-    }: DataProviderInformationAccessors<RealizationSurfaceSettings, RealizationSurfaceData>): [number, number] | null {
+    }: DataProviderInformationAccessors<StatisticalSurfaceSettings, StatisticalSurfaceData>): [number, number] | null {
         const data = getData()?.surfaceData;
         if (!data) {
             return null;
@@ -93,8 +82,10 @@ export class RealizationSurface
     defineDependencies({
         helperDependency,
         availableSettingsUpdater,
+        workbenchSession,
         queryClient,
-    }: DefineDependenciesArgs<RealizationSurfaceSettings>) {
+    }: DefineDependenciesArgs<StatisticalSurfaceSettings>) {
+        availableSettingsUpdater(Setting.STATISTIC_FUNCTION, () => Object.values(SurfaceStatisticFunction_api));
         availableSettingsUpdater(Setting.ENSEMBLE, ({ getGlobalSetting }) => {
             const fieldIdentifier = getGlobalSetting("fieldId");
             const ensembles = getGlobalSetting("ensembles");
@@ -105,21 +96,32 @@ export class RealizationSurface
 
             return ensembleIdents;
         });
-
-        availableSettingsUpdater(Setting.REALIZATION, ({ getLocalSetting, getGlobalSetting }) => {
+        availableSettingsUpdater(Setting.SENSITIVITY, ({ getLocalSetting }) => {
             const ensembleIdent = getLocalSetting(Setting.ENSEMBLE);
-            const realizationFilterFunc = getGlobalSetting("realizationFilterFunction");
 
             if (!ensembleIdent) {
                 return [];
             }
 
-            const realizations = realizationFilterFunc(ensembleIdent);
-
-            return [...realizations];
+            const ensembleSet = workbenchSession.getEnsembleSet();
+            const currentEnsemble = ensembleSet.findEnsemble(ensembleIdent);
+            const sensitivities = currentEnsemble?.getSensitivities()?.getSensitivityArr() ?? [];
+            if (sensitivities.length === 0) {
+                return [];
+            }
+            const availableSensitivityPairs: SensitivityNameCasePair[] = [];
+            sensitivities.map((sensitivity) =>
+                sensitivity.cases.map((sensitivityCase) => {
+                    availableSensitivityPairs.push({
+                        sensitivityName: sensitivity.name,
+                        sensitivityCase: sensitivityCase.name,
+                    });
+                }),
+            );
+            return availableSensitivityPairs;
         });
 
-        const realizationSurfaceMetadataDep = helperDependency(async ({ getLocalSetting, abortSignal }) => {
+        const surfaceMetadataDep = helperDependency(async ({ getLocalSetting, abortSignal }) => {
             const ensembleIdent = getLocalSetting(Setting.ENSEMBLE);
 
             if (!ensembleIdent) {
@@ -138,7 +140,7 @@ export class RealizationSurface
         });
 
         availableSettingsUpdater(Setting.ATTRIBUTE, ({ getHelperDependency }) => {
-            const data = getHelperDependency(realizationSurfaceMetadataDep);
+            const data = getHelperDependency(surfaceMetadataDep);
 
             if (!data) {
                 return [];
@@ -150,10 +152,9 @@ export class RealizationSurface
 
             return availableAttributes;
         });
-
         availableSettingsUpdater(Setting.SURFACE_NAME, ({ getHelperDependency, getLocalSetting }) => {
             const attribute = getLocalSetting(Setting.ATTRIBUTE);
-            const data = getHelperDependency(realizationSurfaceMetadataDep);
+            const data = getHelperDependency(surfaceMetadataDep);
 
             if (!attribute || !data) {
                 return [];
@@ -173,7 +174,7 @@ export class RealizationSurface
         availableSettingsUpdater(Setting.TIME_OR_INTERVAL, ({ getLocalSetting, getHelperDependency }) => {
             const attribute = getLocalSetting(Setting.ATTRIBUTE);
             const surfaceName = getLocalSetting(Setting.SURFACE_NAME);
-            const data = getHelperDependency(realizationSurfaceMetadataDep);
+            const data = getHelperDependency(surfaceMetadataDep);
 
             if (!attribute || !surfaceName || !data) {
                 return [];
@@ -206,51 +207,81 @@ export class RealizationSurface
 
     fetchData({
         getSetting,
+        getWorkbenchSession,
         registerQueryKey,
         queryClient,
-    }: FetchDataParams<RealizationSurfaceSettings, RealizationSurfaceData>): Promise<RealizationSurfaceData> {
+    }: FetchDataParams<StatisticalSurfaceSettings, StatisticalSurfaceData>): Promise<StatisticalSurfaceData> {
         let surfaceAddress: FullSurfaceAddress | null = null;
         const addrBuilder = new SurfaceAddressBuilder();
 
         const ensembleIdent = getSetting(Setting.ENSEMBLE);
-        const realizationNum = getSetting(Setting.REALIZATION);
         const surfaceName = getSetting(Setting.SURFACE_NAME);
         const attribute = getSetting(Setting.ATTRIBUTE);
         const timeOrInterval = getSetting(Setting.TIME_OR_INTERVAL);
+        const statisticFunction = getSetting(Setting.STATISTIC_FUNCTION);
+        const sensitivityNameCasePair = getSetting(Setting.SENSITIVITY);
 
-        if (ensembleIdent && surfaceName && attribute && realizationNum !== null) {
+        const workbenchSession = getWorkbenchSession();
+
+        if (ensembleIdent && surfaceName && attribute) {
             addrBuilder.withEnsembleIdent(ensembleIdent);
             addrBuilder.withName(surfaceName);
             addrBuilder.withAttribute(attribute);
-            addrBuilder.withRealization(realizationNum);
+
+            // Get filtered realizations from workbench
+            let filteredRealizations = workbenchSession
+                .getRealizationFilterSet()
+                .getRealizationFilterForEnsembleIdent(ensembleIdent)
+                .getFilteredRealizations();
+            const currentEnsemble = workbenchSession.getEnsembleSet().findEnsemble(ensembleIdent);
+
+            // If sensitivity is set, filter realizations further to only include the realizations that are in the sensitivity
+            if (sensitivityNameCasePair) {
+                const sensitivity = currentEnsemble
+                    ?.getSensitivities()
+                    ?.getCaseByName(sensitivityNameCasePair.sensitivityName, sensitivityNameCasePair.sensitivityCase);
+
+                const sensitivityRealizations = sensitivity?.realizations ?? [];
+
+                filteredRealizations = filteredRealizations.filter((realization) =>
+                    sensitivityRealizations.includes(realization),
+                );
+            }
+
+            // If realizations are filtered, update the address
+            const allRealizations = currentEnsemble?.getRealizations() ?? [];
+            if (!isEqual([...allRealizations], [...filteredRealizations])) {
+                addrBuilder.withStatisticRealizations([...filteredRealizations]);
+            }
 
             if (timeOrInterval !== SurfaceTimeType_api.NO_TIME) {
                 addrBuilder.withTimeOrInterval(timeOrInterval);
             }
 
-            surfaceAddress = addrBuilder.buildRealizationAddress();
-        } else {
-            console.debug("RealizationSurfaceLayer: Missing required settings for fetching data");
+            if (statisticFunction) {
+                addrBuilder.withStatisticFunction(statisticFunction);
+            }
+            surfaceAddress = addrBuilder.buildStatisticalAddress();
         }
 
         const surfAddrStr = surfaceAddress ? encodeSurfAddrStr(surfaceAddress) : null;
 
-        const queryKey = ["getSurfaceData", surfAddrStr, null, "png"];
+        const queryOptions = getSurfaceDataOptions({
+            query: {
+                surf_addr_str: surfAddrStr ?? "",
+                data_format: this._dataFormat,
+                resample_to_def_str: null,
+            },
+        });
 
-        registerQueryKey(queryKey);
+        registerQueryKey(queryOptions.queryKey);
 
         const promise = queryClient
             .fetchQuery({
-                ...getSurfaceDataOptions({
-                    query: {
-                        surf_addr_str: surfAddrStr ?? "",
-                        data_format: this._dataFormat,
-                        resample_to_def_str: null,
-                    },
-                }),
+                ...queryOptions,
             })
             .then((data) => ({ format: this._dataFormat, surfaceData: transformSurfaceData(data) }));
 
-        return promise as Promise<RealizationSurfaceData>;
+        return promise as Promise<StatisticalSurfaceData>;
     }
 }
